+++ src/components/BatchTransactionsTracker/BatchTransactionsTracker.tsx
@@ -1,9 +1,0 @@
-import {
-  BatchTransactionsTrackerProps,
-  useBatchTransactionsTracker
-} from 'hooks/transactions/batch/tracker/useBatchTransactionsTracker';
-
-export function BatchTransactionsTracker(props: BatchTransactionsTrackerProps) {
-  useBatchTransactionsTracker(props);
-  return null;
-}