+++ src/components/TransactionsTracker/TransactionTracker.tsx
@@ -1,11 +1,9 @@
-import {
-  useTransactionsTracker,
-  TransactionsTrackerType
-} from './useTransactionsTracker';
+import { useBatchTransactionsTracker } from 'hooks/transactions/batch/tracker/useBatchTransactionsTracker';
+import { TransactionsTrackerType } from 'types/transactionsTracker.types';
+import { useTransactionsTracker } from './useTransactionsTracker';
 
-export function TransactionsTracker({
-  getTransactionsByHash
-}: TransactionsTrackerType) {
-  useTransactionsTracker({ getTransactionsByHash });
+export function TransactionsTracker(props: TransactionsTrackerType) {
+  useTransactionsTracker(props);
+  useBatchTransactionsTracker(props);
   return null;
 }