+++ src/wrappers/DappProvider/CustomComponents.tsx
@@ -1,16 +1,9 @@
 import React from 'react';
-import { BatchTransactionsSender } from 'components/BatchTransactionsSender';
-import { BatchTransactionsTracker } from 'components/BatchTransactionsTracker/BatchTransactionsTracker';
 import { LogoutListener } from 'components/LogoutListener';
-import {
-  TransactionSender,
-  TransactionSenderType
-} from 'components/TransactionSender';
-import {
-  TransactionsTracker,
-  TransactionsTrackerType
-} from 'components/TransactionsTracker';
-import { BatchTransactionsTrackerProps } from 'hooks/transactions/batch/tracker/useBatchTransactionsTracker';
+import { TransactionSender } from 'components/TransactionSender';
+import { TransactionSenderType } from 'components/TransactionSender/types/transactionSender.types';
+import { TransactionsTracker } from 'components/TransactionsTracker';
+import { TransactionsTrackerType } from 'types/transactionsTracker.types';
 
 export interface CustomComponentsType {
   transactionSender?: {
@@ -18,53 +11,26 @@ export interface CustomComponentsType {
     props?: TransactionSenderType;
   };
   transactionTracker?: {
-    component: typeof TransactionsTracker;
+    component?: typeof TransactionsTracker;
     props?: TransactionsTrackerType;
-  };
-  batchTransactionsSender?: {
-    component: typeof BatchTransactionsSender;
-    props?: object;
-  };
-  batchTransactionsTracker?: {
-    component: typeof BatchTransactionsTracker;
-    props?: BatchTransactionsTrackerProps;
   };
 }
 
 export function CustomComponents({
-  customComponents,
-  enableBatchTransactions
+  customComponents
 }: {
   customComponents?: CustomComponentsType;
-  enableBatchTransactions?: boolean;
 }) {
   const transactionSender = customComponents?.transactionSender;
   const transactionTracker = customComponents?.transactionTracker;
-  const batchTransactionsSender = customComponents?.batchTransactionsSender;
-  const batchTransactionsTracker = customComponents?.batchTransactionsTracker;
 
   const TxSender = transactionSender?.component ?? TransactionSender;
   const TxTracker = transactionTracker?.component ?? TransactionsTracker;
-  const BatchTxsSender =
-    batchTransactionsSender?.component ?? BatchTransactionsSender;
-  const BatchTxsTracker =
-    batchTransactionsTracker?.component ?? BatchTransactionsTracker;
 
   return (
     <>
-      {!Boolean(enableBatchTransactions) && (
-        <>
-          <TxSender {...transactionSender?.props} />
-          <TxTracker {...transactionTracker?.props} />
-        </>
-      )}
-
-      {Boolean(enableBatchTransactions) && (
-        <>
-          <BatchTxsSender {...batchTransactionsSender?.props} />
-          <BatchTxsTracker {...batchTransactionsTracker?.props} />
-        </>
-      )}
+      <TxSender {...transactionSender?.props} />
+      <TxTracker {...transactionTracker?.props} />
 
       <LogoutListener />
     </>